Codechange: Move classes to bukkit subpackage
