+++ src/main/java/de/sean/blockprot/bukkit/config/BlockProtConfig.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with BlockProt.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.sean.blockprot.config;
+package de.sean.blockprot.bukkit.config;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
