+++ src/main/java/de/sean/blockprot/bukkit/listeners/PistonEventListener.java
@@ -17,7 +17,7 @@
  */
 package de.sean.blockprot.bukkit.listeners;
 
-import de.sean.blockprot.BlockProt;
+import de.sean.blockprot.bukkit.BlockProt;
 import de.sean.blockprot.bukkit.nbt.BlockNBTHandler;
 import org.bukkit.block.Block;
 import org.bukkit.event.EventHandler;
