+++ src/main/kotlin/de/sean/blockprot/bukkit/TranslationKey.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with BlockProt.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.sean.blockprot
+package de.sean.blockprot.bukkit
 
 import java.util.*
